Fix bug in Vertex packing
git-svn-id: http://svn.cern.ch/guest/lhcb/LHCb/trunk@86480 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: cc5b259a9ac7463c32c03a0f640b9acf577e1d2b

+++ LHCb/Event/PackedEvent/src/PackedParticle.cpp
@@ -250,7 +250,7 @@ void ParticlePacker::unpack( const PackedDataVector & pparts,
 
       // daughters
       
-      for ( int iiD = ppart.firstDaughter; iiD < ppart.lastDaughter; ++iiD )
+      for ( unsigned short int iiD = ppart.firstDaughter; iiD < ppart.lastDaughter; ++iiD )
       {
         const int & iD = pparts.daughters()[iiD];
         int hintID(0), key(0);
